apps/hbys-medical-device/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
pyhumps>=3.8.0
pyserial>=3.5
pyserial-asyncio>=0.6
aiohttp>=3.9.0
python-hl7>=0.4.5
